requirements.txt
flask
flask-cors
requests
beautifulsoup4
pandas
numpy
scikit-learn
nltk
sentence-transformers
faiss-cpu
xgboost
whois
tldextract
joblib
openai
python-dotenv
chromadb
nltk
urllib3
